Place structures in corresponding files.
